Created the ProposalAnalysis Notebook to begin the Analysis.
General Description of the code: This is the code used to begin the analysis of partisan counterporposals for the 2013 and 2017 redistricting process.

Objectives: i. create a new dataframe named ProposalAnalysis.df that combines specific variables of the the proposals.df, RulingParty.df and StateDat.df; ii. uses the ProposalAnalysis.df to build crosstabs, run bivariate analyisis (chi-sq), and/or multivariate analyisis